smallint converted in int

+++ server/db/db_template.sql
@@ -3,7 +3,7 @@ CREATE DATABASE IF NOT EXISTS `fleamaster`;
 
 -- user table
 CREATE TABLE IF NOT EXISTS `user` (
-    'userid' smallint auto_increment primary key,
+    'userid' int auto_increment primary key,
     'firstname' varchar(30),
     'lastname' varchar(30),
     'gender' varchar(4),
@@ -15,8 +15,8 @@ CREATE TABLE IF NOT EXISTS `user` (
 );
 
 Create table 'shoppingCart'(
-  'userid' smallint,
-  'articleid' smallint,
+  'userid' int,
+  'articleid' int,
   'totalprice' float,
   foreign key userid references user(userid),
   foreign key articleid references article(articleid),
@@ -24,17 +24,17 @@ Create table 'shoppingCart'(
 );
 
 Create table 'article'(
-  'articleid' smallint primary key,
-  'provider' smallint,
+  'articleid' int primary key,
+  'provider' int,
   'articlename' varchar(50),
   'description' varchar(255),
-  'categoryid' smallint,
+  'categoryid' int,
   'price' float,
   foreign key categoryid references article(categoryid),
   foreign key provider references user(userid)
 );
 
 CREATE table 'category'(
-  'categoryid' smallint primary key,
+  'categoryid' int primary key,
   'categoryname' varchar(255)
 );
